+++ src/consts.lua
@@ -81,7 +81,7 @@ Consts.DESTROY_EXCESS_SOUL_SHARDS_MAX = 28
 Consts.DESTROY_EXCESS_SOUL_SHARDS_STEP = 1
 Consts.SOUL_SHARD_ITEM_ID = 6265
 
--- destroy.autoSlider
+-- destroy.autoOpen, destroy.autoStart
 Consts.DESTROY_AUTO_SLIDER_MIN = 0
 Consts.DESTROY_AUTO_SLIDER_MAX = 16
 Consts.DESTROY_AUTO_SLIDER_STEP = 1