+++ src/ui/groups/destroy.lua
@@ -29,23 +29,53 @@ function Destroy:AddGeneral(parent)
   Widgets:CheckBoxSlider({
     parent = parent,
     checkBox = {
-      label = L.AUTO_OPEN_DESTROY_TEXT,
+      label = L.AUTO_OPEN_TEXT,
       tooltip = L.AUTO_OPEN_DESTROY_TOOLTIP,
-      get = function() return DB.Profile.destroy.auto end,
-      set = function(value) DB.Profile.destroy.auto = value end
+      get = function() return DB.Profile.destroy.autoOpen.enabled end,
+      set = function(value) DB.Profile.destroy.autoOpen.enabled = value end
     },
     slider = {
-      label = L.AUTO_OPEN_DESTROY_SLIDER_LABEL,
-      tooltip = L.AUTO_OPEN_DESTROY_SLIDER_TOOLTIP:format(Consts.DESTROY_AUTO_SLIDER_MIN),
-      value = DB.Profile.destroy.autoSlider,
+      label = L.THRESHOLD_TEXT,
+      tooltip = L.AUTO_OPTION_THRESHOLD_TOOLTIP:format(
+        "|cFFFFD100"  .. L.AUTO_OPEN_TEXT .. "|r",
+        "|cFFFFD100"  .. L.AUTO_OPEN_TEXT .. "|r"
+      ),
+      value = DB.Profile.destroy.autoOpen.value,
       min = Consts.DESTROY_AUTO_SLIDER_MIN,
       max = Consts.DESTROY_AUTO_SLIDER_MAX,
       step = Consts.DESTROY_AUTO_SLIDER_STEP,
       onValueChanged = function(_, event, value)
-        DB.Profile.destroy.autoSlider = value
+        DB.Profile.destroy.autoOpen.value = value
       end
     }
   })
+
+  -- Auto Start
+  if Addon.IS_CLASSIC then
+    Widgets:CheckBoxSlider({
+      parent = parent,
+      checkBox = {
+        label = L.AUTO_START_TEXT,
+        tooltip = L.AUTO_START_DESTROY_TOOLTIP,
+        get = function() return DB.Profile.destroy.autoStart.enabled end,
+        set = function(value) DB.Profile.destroy.autoStart.enabled = value end
+      },
+      slider = {
+        label = L.THRESHOLD_TEXT,
+        tooltip = L.AUTO_OPTION_THRESHOLD_TOOLTIP:format(
+          "|cFFFFD100"  .. L.AUTO_START_TEXT .. "|r",
+          "|cFFFFD100"  .. L.AUTO_START_TEXT .. "|r"
+        ),
+        value = DB.Profile.destroy.autoStart.value,
+        min = Consts.DESTROY_AUTO_SLIDER_MIN,
+        max = Consts.DESTROY_AUTO_SLIDER_MAX,
+        step = Consts.DESTROY_AUTO_SLIDER_STEP,
+        onValueChanged = function(_, event, value)
+          DB.Profile.destroy.autoStart.value = value
+        end
+      }
+    })
+  end
 
   -- Below Price
   Widgets:CheckBoxSlider({