+++ CHANGELOG.md
@@ -1,5 +1,13 @@
-# 9.0.8
+# 9.0.9
 
-## Changes
-
-- Locale updates
+- **Right-Clicking** the minimap icon will now either:
+  - Destroy Next Item (Retail)
+  - Start Destroying (Classic)
+- Re-added `Destroy > General > Auto Start` option for Classic (See [#87](https://github.com/moody/Dejunk/issues/87))
+- Re-added `Start Selling` keybind (See [#89](https://github.com/moody/Dejunk/issues/89))
+- Re-added `Start Destroying` keybind (Classic only)
+- Added commands:
+  - `/dejunk sell start` - starts the selling process
+  - `/dejunk sell next` - sells the next item
+  - `/dejunk destroy start` - starts the destroying process (Classic only)
+  - `/dejunk destroy next` - destroys the next item