+++ HISTORY.md
@@ -1,3 +1,9 @@
+# 9.0.8
+
+## Changes
+
+- Locale updates
+
 # 9.0.7
 
 ## Changes